Add Facebook APIs (the ID and the SECRET) to project.
Save state.

Add views for various signin pages for artist and users.

Add Google auth ID and Secret and attempt to persist data through local signup to be rendered on multiple views at a time.

There are several approaches we can take here. Have one controller that controlles a the view logic for a subset of views. Or, data can be rendered on each view with a single controller per view. Lastly, a service of some type can be set up to facilitate the persistance of data across each view or be injected across each controller.